pyproject.toml
[project]
name = "evm-backer"
version = "0.1.0"
description = "EVM Ledger Registrar Backer for KERI — anchors KERI key events on Ethereum"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.14.0"
keywords = ["keri", "ethereum", "evm", "zk", "ed25519", "vlei"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Topic :: Security :: Cryptography",
]
dependencies = [
    "keri>=1.3.4",
    "web3>=7.14.1",
    "eth-account>=0.13.0",
    "pynacl>=1.5.0",
]

[project.scripts]
evm-backer = "evm_backer.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/evm_backer"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
